+++ src/model/bullet/HeavyBullet.java
@@ -4,11 +4,11 @@ import config.ImageLoader;
 import model.Shooter;
 
 public class HeavyBullet extends Bullet {
-    public HeavyBullet(double x, double y, double angle, Shooter shooter) {
+    public HeavyBullet(float x, float y, float angle, Shooter shooter) {
         super(x, y, ImageLoader.heavyBullet(), 50, 50, angle, shooter);
         this.damage = 20;
         this.cost = 50;
-        this.speed = 10.0d;
+        this.speed = 10.0f;
         live();
     }
 }
