+++ src/utils/Helpers.java
@@ -9,15 +9,15 @@ public class Helpers {
 
 
     public static boolean hasIntersect(Unit u1, Unit u2) {
-        double left1 = u1.x - u1.modelWidth / 2.0;
-        double right1 = u1.x + u1.modelWidth / 2.0;
-        double top1 = u1.y - u1.modelHeight / 2.0;
-        double bottom1 = u1.y + u1.modelHeight / 2.0;
+        float left1 = (float) (u1.x - u1.modelWidth / 2.0);
+        float right1 = (float) (u1.x + u1.modelWidth / 2.0);
+        float top1 = (float) (u1.y - u1.modelHeight / 2.0);
+        float bottom1 = (float) (u1.y + u1.modelHeight / 2.0);
 
-        double left2 = u2.x - u2.modelWidth / 2.0;
-        double right2 = u2.x + u2.modelWidth / 2.0;
-        double top2 = u2.y - u2.modelHeight / 2.0;
-        double bottom2 = u2.y + u2.modelHeight / 2.0;
+        float left2 = (float) (u2.x - u2.modelWidth / 2.0);
+        float right2 = (float) (u2.x + u2.modelWidth / 2.0);
+        float top2 = (float) (u2.y - u2.modelHeight / 2.0);
+        float bottom2 = (float) (u2.y + u2.modelHeight / 2.0);
 
         return right1 > left2 &&
                 left1 < right2 &&
